Add eslint and apply fixes
