+++ .eslintrc.json
@@ -1,0 +1,36 @@
+{
+  "env": {
+    "es2020": true,
+    "mocha": true
+  },
+  "extends": [
+    "plugin:vue/essential",
+    "plugin:mocha/recommended",
+    "airbnb-base"
+  ],
+  "parserOptions": {
+    "ecmaVersion": 11,
+    "sourceType": "module"
+  },
+  "plugins": [
+    "chai-friendly",
+    "mocha",
+    "vue"
+  ],
+  "rules": {
+    "arrow-parens": ["error", "as-needed"],
+    "comma-dangle": ["error", "never"],
+    "func-names": "off",
+    "prefer-arrow-callback": "off",
+    "import/extensions": ["error", "never"],
+    "no-unused-expressions": "off",
+    "chai-friendly/no-unused-expressions": 2
+  },
+  "settings": {
+    "import/resolver": {
+      "node": {
+        "extensions": [".js", ".css", ".vue"]
+      }
+    }
+  }
+}
