+++ .github/workflows/run-tests.yml
@@ -65,3 +65,4 @@ jobs:
       - run: yarn install --frozen-lockfile
       - run: yarn prod
       - run: yarn test
+      - run: yarn lint
