+++ tests/JavaScript/factory.js
@@ -1,37 +1,27 @@
-export const languageFactory = props => {
-  return {
-    name: 'Test Language',
-    algo_code: 'TL',
-    group: {},
-    ...props
-  };
-};
+export const slotFactory = props => ({
+  abv: 'SLOT',
+  ...props
+});
 
-export const groupFactory = props => {
-  return {
-    name: 'Test Group',
-    ...props
-  };
-};
+export const glossFactory = props => ({
+  abv: 'GLOSS',
+  ...props
+});
 
-export const morphemeFactory = props => {
-  return {
-    slot: slotFactory(),
-    glosses: [glossFactory()],
-    ...props
-  };
-};
+export const languageFactory = props => ({
+  name: 'Test Language',
+  algo_code: 'TL',
+  group: {},
+  ...props
+});
 
-export const slotFactory = props => {
-  return {
-    abv: 'SLOT',
-    ...props
-  }
-};
+export const groupFactory = props => ({
+  name: 'Test Group',
+  ...props
+});
 
-export const glossFactory = props => {
-  return {
-    abv: 'GLOSS',
-    ...props
-  }
-};
+export const morphemeFactory = props => ({
+  slot: slotFactory(),
+  glosses: [glossFactory()],
+  ...props
+});
