+++ tests/JavaScript/setup.js
@@ -2,4 +2,5 @@ require('jsdom-global')();
 
 const chai = require('chai');
 const chaiDom = require('chai-dom');
+
 chai.use(chaiDom);
